+++ src/loadImages.ts
@@ -1,6 +1,6 @@
 import noop from './noop';
 
-/*
+/**
  * Async wrapper for image.onload.
  * @param {string} src Image src
  * @param {VoidFunction} onLoaded Callback on image success load end
@@ -19,7 +19,7 @@ export const loadImage = async (
     curImage.onerror = () => resolve();
   });
 
-/*
+/**
  * Method for async load of multiple images
  * @param {string[]} images Images src array
  * @param {VoidFunction} onNextLoaded Callback on every image success load end